Save original command, and add tag cli argument
The original command used to start the benchmark is now saved to the
metadata of the run.
The --tag argument can be used to add a custom tag with which to
identify your results. This tag will be saved as metadata, and (next up)
also be uploaded to the openml runs.