Improve version information for flow
The description is updated to include a more specific reference to which
version of the benchmark framework is used. Similar for dependencies.
I omitted branch information as this should not be necessary with
tag/commithash and did not want to clutter the description.
On my local setup git information can not be retrieved, so I opted to
simply file "NA" similar to how it is reported in the results.csv file.