Show constant columns separately from the results
With more columns added over time to the results table, the printed
table would generally wrap long lines. This makes the table very hard to
read. At the same time a lot of the information is redundant. This table
separately prints some constant information so the variable data in the
table can be displayed neatly on more screens.