Make use of the 'robot' retry_policy
Currently I see no reason to allow the script to be ran with 'human'
retry policy. The retry policy only affects runs on poor connection or
when the server is in distress. The 'robot' retry policy still retries
relatively quickly in the beginning, so the distinction to 'human' is
not that noticable, but the total numbers of retries allows should make
it more stable.