Prevent downloading qualities
I'm still advocating to make lazy loading the default in openml-python,
but until then we can use this to avoid downloading qualities
altogether. It's not a particularly expensive call (~50ms) on my local
machine, but it the OpenML server panics quite easily so reducing the
amount of queries hopeful helps keep the server stable.