Fix mismatch between arff header and data
Due to the conversion that openml-python does from a categorical arff
column to a boolean pandas column, true/false categoricals are actually
written as True/False (note capitalization) in the data but the features
still describe it as true/false. This leads to a mismatch in the ARFF
header and the actual data, which caused issues for R frameworks.