Add parameter to set tune time, rest is for fit
The previous iteration where the final fit was treated as an equivalent
budget to any other optimization sometimes left too little time to train
the final forest, in particular when the last fit took longer than
expected. This would often lead to very small forests for the final
model. The new system guarantees roughly 10% of budget for the final
forest, guaranteeing a better final fit.