Provide a warning and return early if no process output is detected
This avoids potentially crashing if the logging is configured incorrectly.
In the future, we should expand this to first check how logging is
configured in order to see whether or not the issue should be reported
and possibly give a more detailed warning if it is likely the cause
of an error.