Rewrite to avoid tuples in dataframe/csv
We remove the the `metric` column is now updated to be `optimization_metrics`,
which is a character-separated field listing each metric.
We removed the result column, since it would otherwise contain tuples,
and the data is duplicate with the individual metric columns
(except for the convenience of having an 'higher is better' column,
which we now lose).