Measure inference time (#532)
Add the option to measure inference time (disabled by default) for most frameworks.
For those frameworks, inference time is measured capturing both the data loading and the inference.
This is done to make things more equal between the different frameworks (as some _need_ to read the file if they don't operator in Python). Inference time is measured multiple times for different batch sizes (configurable). By default, the median is reported in the results file (as it is less sensitive to e.g., cold-starts) but all measured inference times are stored in the predictions folder of a run.
For Python frameworks, inference time for in-memory single row predictions is also measured.