Set reasonable defaults for greater time budgets
NAML isnt designed to run for a long time, so the defaults of 20
iterations or 10 second execution timeout is not useful when
timeout is an hour. Based on a conversation with @fmohr we set
more reasonable defaults, leading to behavior which mostly
should make use of the runtime constraints.
I added an easy option to revert to normal defaults by
setting `_use_default_time_and_iterations` as a configuration
option, which will ignore any time constraints or dynamic
defaults derived from it.