Fix FLAML installation problems, support version >=2 (#626)
FLAML was fixed to <2, and would install an incompatible xgboost
version. This patch makes it so that newer versions of FLAML
can be installed. Additionally, when installing an older version
a compatible version of xgboost is installed.
Finally, it adds a definition of flaml:latest in line with the
other framework definitions.