Switch to functools for caching (#700)
* Switch to functools implementation which introduces 'bug'

The bug is that now objects of the same class can not have their
cache cleared independently. From the usage, I do not believe that
there would otherwise be two different instances of the same class
anyway, though I may be wrong (thinking datasplit). Will need
to investigate more tomorrow.

In any case, the plan is to move to a more friendly implementation
anyway where those classes that do need cache eviction simply add
a two-liner for that.

* Use cached_property instead of caching a property manually

* Properly clear properties and also functions

Still has the issue that cache for all instances of the class are
cleared for functions (but not for properties).

* Use functools for caching